MINOR: Pin Rust 1.93.0 (MSRV), switch reqwest to native roots, harden CI/lints
- Centralized shared package metadata + dependency versions
- Removed unused cargo dependencies; fixed machete warnings
- Linting fixes
- Improved CI workflow
- Swapped to native-roots from bundled CA roots
- Pinned rust-version; MSRV 1.93.0

+++ crates/obzenflow_runtime_services/src/runtime_resource_limits.rs
@@ -121,8 +121,8 @@ fn get_nofile_limit() -> io::Result<NofileLimit> {
         }
 
         Ok(NofileLimit {
-            soft: current.rlim_cur as u64,
-            hard: current.rlim_max as u64,
+            soft: current.rlim_cur,
+            hard: current.rlim_max,
         })
     }
 }
